requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
numpy==1.24.4
packaging==23.1
pandas==2.0.3
pluggy==1.3.0
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tomli==2.0.1
tzdata==2023.3
urllib3==2.0.5
